Fix panic in ProcessState.Stop()
The methods ('Terminate', 'Kill', ...) of the internally used *gexec.Session
do not check, if there is actually a *os.Process, they just blindly call
methods (most notably 'Signal') on in.

This bug should be fixed in gexec -- gexec itself should not panic. But
as a hotfix, this should be good enough.