Guard against Panic of APIServer
In the APIServer, if the `EtcdURL` is not configured, it used to panic on
`Start()`. In the arguments template, we called the method `String()` on
the `EtcdURL` even if it was `nil`.

This is now checked against and a proper error is ommited in case
`EtcdURL` is `nil`.

Additionally, we also guard in the default argument templates against
issues like that. We should never run into issues here, as the
defaulting should make it impossible to for our default templates to
call methods on `nil`. However, we want to be good role models and do
out best to guard against error.