Change APIServer startup message
We found that the APIServer was not fully up when it logged "Serving
(in)securely on ...". We tried to find a different log line we can wait
for to know when the APIServer is fully up.

The autoregister controller came in with
https://github.com/kubernetes/kubernetes/pull/42732. It seems to be the
last thing to come up before the server is ready to go.