Fix some issues in types.ReconcileKey
While trying to determine what key to emit for an element without a
namespace, I looked at types.ReconcileKey which would emit "/name" which
doesn't appear to agree with the rest of the logic in kubebuilder. This
fixes types.ReconcileKey to emit the correct string value if namespace
is empty as well as adds a method to construct the object from a string.