Strip glog from the event recorder
The event recorded was using glog to log, which it should not have been.
Plus, the event information is inherently structured, so it's a win to
provide structured logging of the event.  We probably want to have the
recorder have a unique logger per controller (or consumer in general) --
right now, it's just using the glog manager logger.