External clusters in envtest
This allows you to use an external cluster with envtest.  In the
case where `UseExistingCluster` is set to true, envtest will skip
initializing the control plane, and if config is not otherwise set,
it will populate the config from local kubeconfigs or in-cluster config.

(This also shuffles around the types in envtest without breaking compat
to avoid an import loop)