WIP adds Context support to manager in place of stop channels.
This is backward-incompatible. The manager accepts Context instead of a stop
channel, which helps simplify cancellation workflow and is more in line with
the direction related projects are moving.

This pattern can be expanded in a similar way to controllers, the cache, and
so on if people like it.

fixes #103