Add scheme support to fake client
Adds a NewFakeClientWithScheme function that takes a *runtime.Scheme as
its first argument. NewFakeClient is now a wrapper around
NewFakeClientWithScheme, passing it the default scheme.Scheme.

As part of this, the scheme can now be used to derive the correct List
GVK. This eliminates the need to use a metav1.List and
opts.Raw.TypeMeta. Now it's possible to pass in a typed List, same as
with the real client.