Dependency injection context
This implements a context-style lightweight dependency injection.
A Context implements a method that populates a pointer to a value.
Different implementations *may* implement this differently, but the
included implementation functions much like a normal context does, under
the hood (method to add a value, forming a linked list).

Convience methods are provided to inject all dependencies for a given
object, based on the existing method style, or on a struct-tag-based
style for ease of use.