Switch alias.go to new logging pkg layout
This switches alias.go over to the new logging layout.  In order to
avoid alias.go pulling in Zap, we don't put ZapLogger in alias.go any
more.  Practically, this is probably fine.  It means an extra line in
main.go, but it's just one, and allows people to avoid extraneous
dependencies.