Fix admission webhook injection, no mgr, no client
The webhook server logic was not handling injection correctly -- it
implemented injection for a couple of specific fields, rather than
receiving and dealing with an injectfunc.  This fixes that.

As a side effect, it also removes the handle to manager and client from
the webhook server, since those aren't used by the server, and should
thus be injected from the manager.