Flatten webhook types struct
This flattens down the webhook package structure, removing dedicated
`types` packages in favor of having things in the relevant places.  To
do this, the inject interface definitions for admission were also moved
to the admission controller location, but they actually make some amount
of sense living there.

It also renames and restructures a couple of the types for code clarity (e.g.
WebhookTypeMutating to MutatingWebhook) or to follow the convention of
other types in CR (e.g. Making wrapped types from core Kubernetes nested
fields).