Extract out multi-handler support, remove builder
This extracts out the multi-handler support into separate interfaces.
That simplifies the code, makes it easier to test directly, and allows
us to drop the now-extraneous Type field.  This simultaneously removes
the builder (for now) since it doesn't actually make anything simpler.

Along the way, we also refactor some common functionality out into a
helper on reponse itself as opposed to being run by different bits of
the webhook.