Remove path from admission, more response helpers
This removes the path functionality from admission webhooks, since it's
not strictly needed.  Relevant metrics are moved up to the webhook
server, and path is added to register.  This also introduces a couple of
helpers that should make code a bit clearer when writing admission
webhooks (Allowed, Denied, Patched), and an alias file to avoid
importing both webhook and webhook/admission.