Fix admission for decoding for unstructured
Due to a weird interaction between the unstructured decoder (which
demands APIVersion and Kind fields) and the API server (which fails to
set those fields on the embedded object in admission requests), we can't
use the normal decoders, nor can we call json.Unmarshal directly on the
unstructured (since it implements an unmarshaller that calls back into
the decoder).

This detects unstructured objects, and does the right thing for them.