improve builder deprecation messages
It's convention in the Go community that a function's deprecation notice
is separated by a new paragraph from the function's documentation. Tools
like staticcheck determine if a function has been deprecated by checking
the beginning of new paragraphs.

This changelist modifies the deprecation notices in the builder package
to follow this convention, allowing linters like staticcheck,
gometalinter, and golangci-lint to warn users of the package.