Allow setting controller name, fix default
This switches the default controller name in the builder to just
lower(<kind>) which fixes any issues that arrise from it showing up in
the deprecated metrics (which use the workqueue name as part of the
metric name).

It also allows overriding name in case of two controllers reconciling
the same type using the builder.