Switch to interface-based options
These behave more-or-less like functional options, except that a given
option can be applied to more than on option type, so we can deprecate
the 3 different dryrunall variants.

Things with arguments generally become type aliases, with the exception
of MatchingField, which gets deprecated in favor of MatchFields (a type
alias to fields.Set).