Issue #537 - Add apiutil option WithDiscoveryFilter() for NewDynamicRESTMapper()
This new option provides a custom filter that is used to select only the needed
api groups that the controller really needs, instead of the default discover client
that will discover all the groups on the server.

The motivation is since controllers typically use a fixed small number of api groups,
while servers have many dynamic api groups, and in such cases if the controller
has some non-zero latency to the api server, or the server is loaded,
the full groups discovery can take long time (10+ seconds).
The filter can reduce the number of groups drastically.

We might want to consider contributing the filter option to
client-go/discovery client.

Signed-off-by: Guy Margalit <guymguym@gmail.com>