Refactor health probes
This refactors the health probes, bringing the style a bit more in line
with CR.  The health probes themselves are now their own handler, and
you add checks simply by using a map from check name to checker (which
is just a function).

The internal structure should now be more condusive to JSON output as
well.