Don't leak goroutines in controller.New
This quashes a goroutine leak caused by calling controller.New repeatedly without
calling Start.  controller.New was creating a new workqueue, which was
starting goroutines and then expecting to be shut down (by the shutdown
method, which is only called at the end of Start).

To solve that, we move workqueue initialization to controller.Start.
This means that we also move watch starting to controller.Start, but
this seems pretty sensible anyway.