Switch away from using errChan in the manager
Instead of using an error channel in the manager, we just use a mutex
and store one error.  The "waiting" is done by blocking on a channel
which gets closed to signal errors.

This achives the same effect (only return one error) without having the
chance of blocking goroutines; with the old code, goroutines could block
trying to return their results (especially since a few tried to return
nil results), and since runnables can be added after start, there's no
way to appropriately size the channel to avoid this happening (plus no
point, since we only report the first error anyway).

We also only report errors when the occurred, never signaling for errors with
a nil error value.