:bug: Fix builder `For` and `Watches` method conflict
If `For` method has been used, there is no way to specify
own eventhandler. But if `Watches` method is used, an error
of lacking object will happend when no apiType is passed.
If both `For` and `Watches` are used with same apiType,
the later one will be ignored as it has been watched.
Just add nil assertion to avoid object type error.