Fix SetControllerReference to consider multiple CRD versions
An error was returned by a call to `SetControllerReference` if the owner
is already set, but with a different apiVersion (eg. `group/v1beta1`
instead of `group/v1`).

This commit fixes it by ignoring the version in the ref comparison. It
still considers the group though. Having an owner with a different
version does not return an error, but the call to
`SetControllerReference` effectively changes the owner to match the
newer apiVersion.