CreateOrPatch
This patch introduces a variation on the controllerutil.CreateOrUpdate
function named CreateOrPatch. Instead of issuing update calls, the new
function uses a patch to perform a more surgical update to the remote
data. Additionally, the implementation relies on logic similar to the
PatchHelper mechanics in the Cluster API util/patch package. The
resource is converted to unstructured data first in order to patch the
resource and any potential status separately.

Two new OperationResult values were added:

    1. OperationResultUpdatedStatus     - the resource and its status
                                          were updated
    2. OperationResultUpdatedStatusOnly - only the status was updated