Allow controllers to be started without adding them to the manager
As far as I can tell the main reason we add controllers to the manager is to
ensure that we've been elected leader before we start any controllers. The
downside of this design is that it's not possible to stop individual
controllers, or remove controllers from the manager.

I believe this commit is the minimum possible change necessary to allow
controllers to be started and stopped on-demand. It allows a controller to be
created, started, and stopped without ever being added to the manager. Any
controller that is started separately from the manager must handle its own
leader election.

https://github.com/kubernetes-sigs/controller-runtime/issues/730

Signed-off-by: Nic Cope <negz@rk0n.org>