Use metav1.Object instead of Object
In order to allow for a broader adoption of ContainsFinalizer, RemoveFinalizer and AddFinalizer and remove a BC break introduced to 0.5.x I switched the method back to the `metav1.Object`.

After reading https://github.com/kubernetes-sigs/controller-runtime/issues/959 and https://github.com/kubernetes-sigs/controller-runtime/pull/962 I'm unable to understand the benefit of requiring `runtime.Object` so I propose to revert the BC break and possible panic on line 285.