Try to avoid event handling leaks
Since we now have the ability shut down the event broadcaster, we can
write mostly goroutine-leak-free event handling setup.  This changes the
default event handling setup to defer the broadcaster initialization the
first time it's used, and then to shut it down once the manager shuts
down.

In the case where a broadcaster is manually specified, it's the
specifier's job to shut down the broadcaster instead.

We'll probably still want to overhaul the whole event system at some
point in the future though.

This also re-enables the tests for leaks, switching them to an
eventually to avoid flakes & reducing the threshold to zero.