Switch to using goleaks to check for leaks
This switches to using the goleaks package to check for leaks, which
should give us a more complete picture of the particular goroutine
that's leaking, and should avoid issues where we leak a goroutine, but
also stop an old one.

This also force-closes keep-alive connections in the leak tests, since
those look like leaks, but will actually time out after 30s (outside the
timescope of the test).