Avoid the redirection to `/healthz/` when calling `/healthz`
By https://github.com/kubernetes-sigs/controller-runtime/pull/1100,
subpaths of `/healthz/` can be handled.
However, this change introduced the redirection from `/healthz`
to `/healthz/` with 301 (Moved Permanently) when accessing `/healthz`
endpoint. (Accessing `/healthz/` works fine without the redirect.)

This is unnecessary overhead in the health checking.

So, this commit enable to access `/healthz` without the redirect with keeping
that subpaths of `/healthz/` are accessible.