:seedling: Expose handler.MapFunc to allow implementers to test mappers
This is a follow-up from a previous change to the handler package that
deprecated the old handler.ToRequestFunc in favor of
EnqueueRequestsFromMapFunc.

While we still want to prefer using EnqueueRequestsFromMapFunc, this
change also exposes the type of MapFunc signature expected. Users can
use this type to test their own mappers in unit tests.

Signed-off-by: Vince Prignano <vincepri@vmware.com>