:warning: Propagate context on Manager.Start(...)
This change reshuffles how the manager accepts and operates on
a context. With this change, the user experience is greatly improved,
users can now use `ctrl.SetupSignalHandler()` to create a context, enrich
it if they want to, and pass it to `manager.Start`.

In addition, this PR changes how the context and stop channel are
handled internally to ensure proper cancellation.

Signed-off-by: Vince Prignano <vincepri@vmware.com>