:warning: Handler: Remove MapObject type and use client.Object directly
We updated the handlers MapObject type to not include a metav1.Object
and a runtime.Object representation of the same object but instead a
client.Object, which is a union of the two above.

The only reason for the MapObject type to exist in the first place
though was to hold these two fields. Now that there is only the
client.Object left, we can use that directly and get rid of the
MapObject type.