Fix use of `UnstructuredList` with unregistered typed
Before this change, passing an `unstructured.UnstructuredList` to
`client.List()` would erroneously delegate to the typed client, causing an error
if the target type isn't registered with the client's scheme, violating the
purpose of using `unstructured`.

This patch fixes the client so that the `unstructured.UnstructuredList` is
delegated to the internal unstructured client.