:bug: Cache bypass should take into account List types
Currently, when we create a new delegating client and we force some
types to be uncached, we can only specify client.Object(s).

When these objects are passed to the delegating client, and their GVK is
matched against the internal map, the check doesn't take into account
that the client might be issuing a List call with a ObjectList instead
of a metav1.Object.

The proposed solution is taken from the discovery rest mapper, which
does the same ~hack. While it's not perfect, probably close to all
generated objects are going to have a list that has the "List" suffix,
that when removed gives back the original GVK.

If there are other solutions, happy to change this hack.

Signed-off-by: Vince Prignano <vincepri@vmware.com>