✨ Move cluster-specifics from Manager into new pkg/cluster
This change is the first step towards implementing the corresponding
[proposal][1]. It essentially consists of moving a bunch if code from
pkg/manager into pkg/cluster, slightly adjusting some tests in
pkg/manager and copying applicable tests from there into pkg/cluster.

It is not yet the full implementation, because the `Manager` will only
correctly start it's own cache and not other caches before starting the
remaining `Runnables`. This behavior matches current behavior when
building something that uses multiple caches and will be fixed in a
follow-up.

[1]: https://github.com/kubernetes-sigs/controller-runtime/blob/master/designs/move-cluster-specific-code-out-of-manager.md