Warn if logger is not explicitly set
Though we have a 30 second fallback, there is no indication that this
is happening or hint to users that they should set a logger.

This path being hit is not ideal: we may still buffer a substantial
amount of log messages on larger clusters, and we aren't getting any
log messages.  We should log something; even though the best we can do
is to just log an unformatted message to stderr.