Envtest secure serving by default, 1.20 flags
This introduces secure serving by default, transparently switching
configs to connect to secure endpoints.  Similarly, the insecure
endpoints are now disabled by default -- they must be explicitly
disabled if you want to use them (note that this will only work on API
servers <=1.19).

This also turns on flags around the tokenrequest endpoint which are
required to be on in 1.20, and have been available for all the versions
that we currently support (back to 1.16).

There's a whole new set of control plane APIs for provisioning users,
and the old "just get a single account" APIs of the internal package are
mostly deprecated.

The default `Environment.Config` is *NOT* deprecated however -- this is
intended to continue working as normal in order to avoid breaking
everyone and to keep things in envtest working easily -- most tests will
be fine with an admin user.

For users that want RBAC, individual users may be provisioned with the
ControlPlane.AddUser method.