:bug: Run unittests in CI
This accidentally got disabled in
cd065bf2f63c3057db307990c21316483fa60ce8 under the assumption that the
CI entrypoint is the Makefile in an attempt to make this configurable.
Since the script is called `test_all.sh`, it is reasonable to assume it
always test verything. If ppl want something else, they can just add
another script/maketarget.