Construct the typed clients we need (#1637)
The pattern of constructing all the clients to get a
kubernetes.Interface, only to extract one or two typed clients has the
side-effect of importing a lot of packages that are otherwise
unneeded.

Avoid doing so by constructing the clients we need.

This isn't sufficient to remove all these extra dependencies (because
of the use of scheme.Scheme as default), but it is a step closer.