🐛  Start the Cache if the Manager has already started (#1681)
* Start the Cache if the Manager has already started

* verify adding cluster after mgr started results in working cache

Add a test that adds a cluster to the manager after the manager has
already started. Verify that the cluster is started and its cache
is sycned.

Added the startClusterAfterManager struct which is a basically just a
hook to verify that the cluster is started.

* remove unnecessary methods

Only GetCache and Start methods are neeed for new test that adds a
cluster to the manage after the manager has already started.