Improve performance of CreateOrPatch
When either of the before/after objects has no status data, the
reflect.DeepEqual call becomes redundant because comparing of
existing to non-existing always returns False.

Switch the logical OR condition to AND. That saves us a single
reflect.DeepEqual() call each time we end up with an object without
status after its mutation func was called (or before that).

Signed-off-by: Bogdan Dobrelya <bdobreli@redhat.com>